first pass splitting discover backend into multiple routes per algorithm
